put sep24 code in a sep24.ts file
